feat(Event-Store): Add `Terminate` as a way of gracefully shutting down
As part of working on the PostgreSQL storage engine, I realised that for most of the engine
implementations, I would need a way to signify that the application itself is shutting down, and a
graceful shutdown is required. This could entail closing connections, draining pools, etc. Basically
perform some good-citizen duties.